+++ js/lib/replacematches.js
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = function multiReplace(str, m) {
-  return str.replace(/\${(\d+)}|\$(\d+)/g, function(tmp, i, j) {
-    return m[(i||j)] || '';
-  }).trim();
-};